format

+++ ethereum/contracts/example_integrations/xMint/Spoke.sol
@@ -46,7 +46,9 @@ contract XmintSpoke is IWormholeReceiver {
     //And then requests delivery from relayer network.
     function purchaseTokens() public payable {
         //Calculate how many tokens will be required to cover transaction fees.
-        uint256 deliveryFeeBuffer = core_relayer.quoteGasDeliveryFee(hub_contract_chain, SAFE_DELIVERY_GAS_CAPTURE, core_relayer.getDefaultRelayProvider());
+        uint256 deliveryFeeBuffer = core_relayer.quoteGasDeliveryFee(
+            hub_contract_chain, SAFE_DELIVERY_GAS_CAPTURE, core_relayer.getDefaultRelayProvider()
+        );
 
         //require that enough funds were paid to cover this transaction and the relay costs
         require(msg.value > deliveryFeeBuffer + core_bridge.messageFee());
@@ -54,7 +56,13 @@ contract XmintSpoke is IWormholeReceiver {
         uint256 bridgeAmount = msg.value - deliveryFeeBuffer - core_bridge.messageFee();
 
         (bool success, bytes memory data) = address(token_bridge).call{value: bridgeAmount + core_bridge.messageFee()}(
-            abi.encodeWithSignature("wrapAndTransferETHWithPayload(unit16,bytes32,uint32,bytes)", hub_contract_chain, hub_contract_address, nonce, abi.encode(core_relayer.toWormholeFormat(msg.sender)))
+            abi.encodeWithSignature(
+                "wrapAndTransferETHWithPayload(unit16,bytes32,uint32,bytes)",
+                hub_contract_chain,
+                hub_contract_address,
+                nonce,
+                abi.encode(core_relayer.toWormholeFormat(msg.sender))
+            )
         );
 
         //Request delivery from the relayer network.
@@ -64,17 +72,25 @@ contract XmintSpoke is IWormholeReceiver {
     //This function receives messages back from the Hub contract and distributes the tokens to the user.
     function receiveWormholeMessages(bytes[] memory vaas, bytes[] memory additionalData) public payable override {
         //Complete the token bridge transfer
-        ITokenBridge.TransferWithPayload memory transferResult = token_bridge.parseTransferWithPayload(token_bridge.completeTransferWithPayload(vaas[0]));
-        require (transferResult.fromAddress == hub_contract_address && core_bridge.parseVM(vaas[0]).emitterChainId == hub_contract_chain);
+        ITokenBridge.TransferWithPayload memory transferResult =
+            token_bridge.parseTransferWithPayload(token_bridge.completeTransferWithPayload(vaas[0]));
+        require(
+            transferResult.fromAddress == hub_contract_address
+                && core_bridge.parseVM(vaas[0]).emitterChainId == hub_contract_chain
+        );
 
         //TODO is the token address the token being transferred, or the origin address?
         ERC20 token = ERC20(core_relayer.fromWormholeFormat(transferResult.tokenAddress));
 
-        token.transfer(core_relayer.fromWormholeFormat(bytesToBytes32(transferResult.payload, 0)), transferResult.amount);
+        token.transfer(
+            core_relayer.fromWormholeFormat(bytesToBytes32(transferResult.payload, 0)), transferResult.amount
+        );
     }
 
     function requestDelivery() internal {
-        uint256 computeBudget = core_relayer.quoteGasDeliveryFee(hub_contract_chain, SAFE_DELIVERY_GAS_CAPTURE, core_relayer.getDefaultRelayProvider());
+        uint256 computeBudget = core_relayer.quoteGasDeliveryFee(
+            hub_contract_chain, SAFE_DELIVERY_GAS_CAPTURE, core_relayer.getDefaultRelayProvider()
+        );
         uint256 applicationBudget = 0;
 
         ICoreRelayer.DeliveryRequest memory request = ICoreRelayer.DeliveryRequest(
@@ -82,20 +98,21 @@ contract XmintSpoke is IWormholeReceiver {
             hub_contract_address, //target address
             hub_contract_address, //refund address, This will be ignored on the target chain because the intent is to perform a forward
             computeBudget, //compute budget
-            applicationBudget, //application budget, not needed in this case. 
+            applicationBudget, //application budget, not needed in this case.
             core_relayer.getDefaultRelayParams() //no overrides
         );
 
-        core_relayer.requestDelivery{value: computeBudget + applicationBudget}(request, nonce, core_relayer.getDefaultRelayProvider());
+        core_relayer.requestDelivery{value: computeBudget + applicationBudget}(
+            request, nonce, core_relayer.getDefaultRelayProvider()
+        );
     }
 
-    function bytesToBytes32(bytes memory b, uint offset) private pure returns (bytes32) {
+    function bytesToBytes32(bytes memory b, uint256 offset) private pure returns (bytes32) {
         bytes32 out;
 
-        for (uint i = 0; i < 32; i++) {
+        for (uint256 i = 0; i < 32; i++) {
             out |= bytes32(b[offset + i] & 0xFF) >> (i * 8);
         }
         return out;
     }
-
 }
